+++ rsocket-core/src/commonMain/kotlin/io/rsocket/kotlin/payload/Payload.kt
@@ -45,3 +45,12 @@ fun Payload(data: ByteArray, metadata: ByteArray? = null): Payload = Payload(
     data = buildPacket { writeFully(data) },
     metadata = metadata?.let { buildPacket { writeFully(it) } }
 )
+
+/**
+ * Wrap data and metadata arrays without copying them.
+ * Changes in input arrays will change payload data, same as reading from payload will change input arrays.
+ */
+fun Payload.Companion.wrap(data: ByteArray, metadata: ByteArray? = null): Payload = Payload(
+    data = ByteReadPacket(data),
+    metadata = metadata?.let { ByteReadPacket(it) }
+)
