+++ rsocket-core/src/commonTest/kotlin/io/rsocket/kotlin/SetupRejectionTest.kt
@@ -16,14 +16,9 @@
 
 package io.rsocket.kotlin
 
-import io.rsocket.kotlin.connection.*
-import io.rsocket.kotlin.core.*
 import io.rsocket.kotlin.error.*
 import io.rsocket.kotlin.frame.*
-import io.rsocket.kotlin.frame.io.*
-import io.rsocket.kotlin.keepalive.*
-import io.rsocket.kotlin.payload.*
-import kotlinx.coroutines.*
+import io.rsocket.kotlin.test.*
 import kotlin.test.*
 
 class SetupRejectionTest : SuspendTest {
