+++ rsocket-core/src/commonTest/kotlin/io/rsocket/kotlin/keepalive/KeepAliveTest.kt
@@ -21,8 +21,7 @@ import io.rsocket.kotlin.*
 import io.rsocket.kotlin.error.*
 import io.rsocket.kotlin.frame.*
 import io.rsocket.kotlin.internal.*
-import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.*
+import io.rsocket.kotlin.test.*
 import kotlin.test.*
 import kotlin.time.*
