+++ rsocket-test/src/commonMain/kotlin/io/rsocket/kotlin/test/TestConnection.kt
@@ -14,14 +14,11 @@
  * limitations under the License.
  */
 
-package io.rsocket.kotlin
+package io.rsocket.kotlin.test
 
 import io.ktor.utils.io.core.*
 import io.rsocket.kotlin.connection.*
 import io.rsocket.kotlin.frame.*
-import kotlinx.coroutines.*
-import kotlinx.coroutines.channels.*
-import kotlinx.coroutines.flow.*
 
 class TestConnection : Connection {
     override val job: Job = Job()
