+++ rsocket-test/src/commonMain/kotlin/io/rsocket/kotlin/test/TestRSocket.kt
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package io.rsocket.kotlin
+package io.rsocket.kotlin.test
 
 import io.ktor.utils.io.core.*
+import io.rsocket.kotlin.*
 import io.rsocket.kotlin.payload.*
-import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.*
 
 class TestRSocket : RSocket {
     override val job: Job = Job()
