+++ rsocket-transport-local/build.gradle.kts
@@ -34,7 +34,7 @@ kotlin {
         }
         val commonTest by getting {
             dependencies {
-                implementation(project(":rsocket-transport-test"))
+                implementation(project(":rsocket-test"))
             }
         }
     }
